Integrate Git commit ID in Gradle build script
This change implements getting the short SHA ID of the current Git commit in the Gradle build script and passing it as a system property to the Java compiler task. This will help tag the specific Git version of the build at compile time for easy tracking and version control.